+++ src/saihubot-cli-skill-balance.js
@@ -5,112 +5,9 @@ import Table from 'ink-table';
 import { t } from 'saihubot-cli-adapter/dist/i18n';
 
 import {getConfig, getNodeURL, xdaiFetch} from './utils';
-import {rpcEthBalance} from './saihubot-cli-skill-chain';
+import {rpcEthBalance, rpcTokenBalance} from './saihubot-cli-skill-chain';
 import {i18nValidator} from './i18n';
-
-export const contractMap = {
-  "0x4Fabb145d64652a948d72533023f6E7A623C7C53": {
-    "name": "Binance USD",
-    "logo": "busd.svg",
-    "erc20": true,
-    "symbol": "BUSD",
-    "decimals": 18
-  },
-  "0x6B175474E89094C44Da98b954EedeAC495271d0F": {
-    "name": "Dai Stablecoin",
-    "logo": "dai.svg",
-    "erc20": true,
-    "symbol": "DAI",
-    "decimals": 18
-  },
-  "0x8E870D67F660D95d5be530380D0eC0bd388289E1": {
-    "name": "PAX Stablecoin",
-    "logo": "pax.svg",
-    "erc20": true,
-    "symbol": "PAX",
-    "decimals": 18
-  },
-  "0x0000000000085d4780B73119b644AE5ecd22b376": {
-    "name": "TrueUSD",
-    "logo": "tusd.png",
-    "erc20": true,
-    "symbol": "TUSD",
-    "decimals": 18
-  },
-  "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48": {
-    "name": "USD Coin",
-    "logo": "usdc.png",
-    "erc20": true,
-    "symbol": "USDC",
-    "decimals": 6
-  },
-  "0xdAC17F958D2ee523a2206206994597C13D831ec7": {
-    "name": "Tether USD",
-    "logo": "tether_usd.png",
-    "erc20": true,
-    "symbol": "USDT",
-    "decimals": 6
-  },
-  "0x39AA39c021dfbaE8faC545936693aC917d5E7563": {
-    "name": "Compound USD Coin",
-    "logo": "ctoken-usdc.svg",
-    "erc20": true,
-    "symbol":"cUSDC",
-    "decimals": 8,
-    "source": "Compound"
-  },
-  "0xf650C3d88D12dB855b8bf7D11Be6C55A4e07dCC9": {
-    "name": "Compound Tether",
-    "logo": "ctoken-usdt.svg",
-    "erc20": true,
-    "symbol":"cUSDT",
-    "decimals": 8,
-    "source": "Compound"
-  },
-  "0x5d3a536E4D6DbD6114cc1Ead35777bAB948E3643": {
-    "name": "Compound Dai",
-    "logo": "ctoken-dai.svg",
-    "erc20": true,
-    "symbol":"cDAI",
-    "decimals": 8,
-    "source": "Compound"
-  },
-  "0x6Ee0f7BB50a54AB5253dA0667B0Dc2ee526C30a8": {
-    "name": "Aave Interest bearing BUSD",
-    "erc20": true,
-    "symbol":"aBUSD",
-    "decimals": 18,
-    "source": "AAVE"
-  },
-  "0xfC1E690f61EFd961294b3e1Ce3313fBD8aa4f85d": {
-    "name": "Aave Interest bearing DAI",
-    "erc20": true,
-    "symbol":"aDAI",
-    "decimals": 18,
-    "source": "AAVE"
-  },
-  "0x4DA9b813057D04BAef4e5800E36083717b4a0341": {
-    "name": "Aave Interest bearing TUSD",
-    "erc20": true,
-    "symbol":"aTUSD",
-    "decimals": 18,
-    "source": "AAVE"
-  },
-  "0x9bA00D6856a4eDF4665BcA2C2309936572473B7E": {
-    "name": "Aave Interest bearing USDC",
-    "erc20": true,
-    "symbol":"aUSDC",
-    "decimals": 6,
-    "source": "AAVE"
-  },
-  "0x71fc860F7D3A592A4a98740e39dB31d25db65ae8": {
-    "name": "Aave Interest bearing USDT",
-    "erc20": true,
-    "symbol":"aUSDT",
-    "decimals": 6,
-    "source": "AAVE"
-  }
-}
+import {tokenMap, xdaiTokenMap} from './token';
 
 const balanceI18n = {
   'en': {
@@ -129,6 +26,54 @@ const balanceI18n = {
   },
   props: ['balance', 'usdt']
 };
+
+const EthBalances = ({addresses, fetch}) => {
+  const [balance, setBalance] = useState([]);
+  if (!addresses) return null;
+  const data = [];
+
+  useEffect(() => {
+    async function fetchEthBalance() {
+      const ethBalances = await getEtherBalances(getNodeURL(), addresses);
+      Object.values(ethBalances).map(val => {
+        if (val === 0n) return;
+        data.push({
+          [t('token', {i18n:balanceI18n})]: 'ETH',
+          [t('balance', {i18n: balanceI18n})]: (Number(val) / 10**18).toFixed(8),
+          [t('source', {i18n: balanceI18n})]: '',
+        });
+      });
+      setBalance([...balance, ...data]);
+    }
+
+    async function fetchTokenBalance() {
+      const tokenBalances = await getTokensBalances(getNodeURL(), addresses, Object.keys(tokenMap));
+        Object.keys(tokenBalances).map(addr =>
+          Object.entries(tokenBalances[addr]).map(([key, val]) => {
+            if (val === 0n) return;
+            const token = tokenMap[key];
+            data.push({
+              Symbol: token.symbol,
+              Balance: (Number(val) / 10**token.decimals).toFixed(4),
+              [t('source', {i18n: balanceI18n})]: token.source || '',
+            });
+          })
+        );
+        setBalance([...balance, ...data]);
+    }
+    if (addresses) {
+      fetchEthBalance();
+      fetchTokenBalance();
+    }
+  }, [addresses, fetch]);
+
+  return balance.length > 0
+  ? (<>
+    <Text>{t('accountBalance', {i18n: i18nValidator})}</Text>
+    <Table data={balance} />
+  </>)
+  : (<Text>{t('query', {i18n: balanceI18n})}</Text>);
+}
 
 // support multiple validators balance by comma (without space)
 const ValidatorBalances = ({validator, fetch}) => {
@@ -177,6 +122,8 @@ const XdaiBalances = ({addresses, fetch}) => {
   const data = [];
   useEffect(() => {
     async function fetchXdaiBalance() {
+      const tokens = Object.keys(xdaiTokenMap);
+
       for (let i = 0; i < addresses.length ; i++) {
         const json = await xdaiFetch(fetch, rpcEthBalance(addresses[i]))
         const val = json.result === 0x0 ? 0 : Number(json.result)/10**18;
@@ -186,6 +133,19 @@ const XdaiBalances = ({addresses, fetch}) => {
             Balance: val,
             [t('source', {i18n: balanceI18n})]: '',
           });
+        }
+
+        for(let j = 0; j < tokens.length; j++) {
+          const currentTokenAddr = tokens[j];
+          const tokenInfo = xdaiTokenMap[currentTokenAddr];
+          const tokenJson = await xdaiFetch(fetch, rpcTokenBalance(addresses[i], currentTokenAddr));
+          if (tokenJson.result !== '0x0000000000000000000000000000000000000000000000000000000000000000') {
+            data.push({
+              Symbol: tokenInfo.symbol,
+              Balance: Number(tokenJson.result) / 10 ** tokenInfo.decimals,
+              [t('source', {i18n: balanceI18n})]: '',
+            });
+          }
         }
       }
       setBalance([...balance, ...data]);
@@ -197,7 +157,6 @@ const XdaiBalances = ({addresses, fetch}) => {
   return balance.length > 0
     ? (<>
       <Text>{t('xdaiBalance', {i18n: i18nValidator})}</Text>
-      <Text>{''}</Text>
       <Table data={balance} />
     </>)
     : (<Text>{t('query', {i18n: i18nValidator})}</Text>);
@@ -206,40 +165,10 @@ const XdaiBalances = ({addresses, fetch}) => {
 // support multiple account balance by comma (without space)
 // also shows validator balances
 const Balances = ({address, fetch}) => {
-  const [balance, setBalance] = useState([]);
   const [validator, setValidator] = useState('');
-  const data = [];
   const addresses = address.indexOf(',') > -1 ? address.split(',').map(addr => addr.trim()) : [address];
+
   useEffect(() => {
-    async function fetchEthBalance() {
-      const ethBalances = await getEtherBalances(getNodeURL(), addresses);
-      Object.values(ethBalances).map(val => {
-        if (val === 0n) return;
-        data.push({
-          [t('token', {i18n:balanceI18n})]: 'ETH',
-          [t('balance', {i18n: balanceI18n})]: (Number(val) / 10**18).toFixed(8),
-          [t('source', {i18n: balanceI18n})]: '',
-        });
-      });
-      setBalance([...balance, ...data]);
-    }
-
-    async function fetchTokenBalance() {
-      const tokenBalances = await getTokensBalances(getNodeURL(), addresses, Object.keys(contractMap));
-        Object.keys(tokenBalances).map(addr =>
-          Object.entries(tokenBalances[addr]).map(([key, val]) => {
-            if (val === 0n) return;
-            const token = contractMap[key];
-            data.push({
-              Symbol: token.symbol,
-              Balance: (Number(val) / 10**token.decimals).toFixed(2),
-              [t('source', {i18n: balanceI18n})]: token.source || '',
-            });
-          })
-        );
-        setBalance([...balance, ...data]);
-    }
-
     async function fetchValidators() {
       for (let i = 0; i < addresses.length ; i++) {
         const json = await fetch(`https://beaconcha.in/api/v1/validator/eth1/${addresses[i]}`)
@@ -255,21 +184,17 @@ const Balances = ({address, fetch}) => {
     }
 
     if (address) {
-      fetchEthBalance();
-      fetchTokenBalance();
       fetchValidators();
     }
   }, [address, fetch]);
 
-  return balance.length > 0
-    ? (<>
-      <Text>{t('accountBalance', {i18n: i18nValidator})}</Text>
-      <Table data={balance} />
-      <Text> </Text>
-      <ValidatorBalances validator={validator} fetch={fetch} />
-      <XdaiBalances addresses={addresses} fetch={fetch} />
-    </>)
-    : (<Text>{t('query', {i18n: balanceI18n})}</Text>);
+  return (<>
+    <EthBalances addresses={addresses} fetch={fetch} />
+    <Text> </Text>
+    <ValidatorBalances validator={validator} fetch={fetch} />
+    <Text> </Text>
+    <XdaiBalances addresses={addresses} fetch={fetch} />
+  </>)
 }
 
 /**